chore(naming): rename non-kebab-case directories to kebab-case
