+++ huaweicloud/services/internal/httpclient-go/signer.go
