+++ scripts/markdown-gen/writer.go
