+++ scripts/read-check-deleted-check/main.go
